Hmm 32bit

+++ src/core/array2dim_api.h
@@ -32,7 +32,7 @@
           unsigned long gt_a2d_i;                                              \
           ARRAY2DIM = gt_malloc(sizeof *ARRAY2DIM * (ROWS));                   \
           (ARRAY2DIM)[0] = gt_malloc(sizeof **ARRAY2DIM * (ROWS) * (COLUMNS)); \
-          for (gt_a2d_i = 1UL; gt_a2d_i < (ROWS); gt_a2d_i++)                  \
+          for (gt_a2d_i = 1UL; gt_a2d_i < (unsigned long) (ROWS); gt_a2d_i++)  \
             (ARRAY2DIM)[gt_a2d_i] = (ARRAY2DIM)[gt_a2d_i-1] + (COLUMNS);       \
         }
 
@@ -47,7 +47,7 @@
           ARRAY2DIM = gt_malloc(sizeof *ARRAY2DIM * (ROWS));                  \
           (ARRAY2DIM)[0] = gt_calloc((size_t) ((ROWS) * (COLUMNS)),           \
             sizeof **ARRAY2DIM);                                              \
-          for (gt_a2d_i = 1UL; gt_a2d_i < (ROWS); gt_a2d_i++)                 \
+          for (gt_a2d_i = 1UL; gt_a2d_i < (unsigned long) (ROWS); gt_a2d_i++) \
             (ARRAY2DIM)[gt_a2d_i] = (ARRAY2DIM)[gt_a2d_i-1] + (COLUMNS);      \
         }
